Lots of boilerplate code
Changed HTML landing page. Will need to fix <span> elements.  Added form
wrappers for actions

Added lots of boilerplate code to start the ball rolling.

Added DAO, Recipe/RecipeCard beans with minor inputs

Created a "resources" package to house the DDL for the SQL DB. Don't
know if this will work or not, but thought it would be worth the try!

Added servlets for create user and login.  no functionality added.
Probably should wait and add working functionality before adjusting ...
:(  Thinking I need to focus on one aspect of the application, and I
think that's getting the DB up and running easily/efficiently.  Probably
need to use the init() function in servlet lifecycle to enable it and
fill it with data.