User authentication added
Added basic user authentication.  Redirects from user signup page to
user authentication page and asks for a code (UUID).  UUID is printed to
console for use in the front-end authentication input box.  Added a few
input validation methods to the UserDAO and called them from the
RegisterUserServlet and UserAuthentication servlet.