Created New JSP File userProfile.jsp
The userProfile.jsp page is used to display the currently logged in user
profile information. It uses the .jsp <% %> features to pull the
information to be displayed to the user from the session object and
place it inside the HTML <inputs>.

It also allows for the deletion of the User account
from the database via the delete button which is handled by the
deleteUserServlet.java

It will also be used to update the user information but currently
that feature is not functional. (despite much effort)