Updated Servlet UpdateUseProfileServlet.java
After fixing my error in login.jsp I was able to uncomment and remove my
testing code and implement my original method of updating the user.

I am able to pull new vales from the <input> fields on login.jsp
pass them to the UserDAO.updateUserDataInDatabase() method and and the
corresponding string values will be updated in the database.

An update success message is also displayed and the userProfile.jsp
page is reloaded.

My submission for Capstone2 is now feature complete!!!!

^_^    ^_^    ^_^