Updated JSP File: header.jsp
Added calls to request object to get current page URI path and save as
String.

One URI is stored I then use the substring method to get the end of this
string and store it in String pageName.

So if URI = /PerfectMeal/index.jsp

pageName will = index.jsp

Found reference for this method at:
https://stackoverflow.com/questions/11130620/how-to-retrieve-the-current-page-name-in-jsp

Added some system.out.prinln and out.println calls to view this as I
work on developing other features.