Updated Servlet File: RegisterUserServlet
Replaced object of type UserDAO with RegisterUserProxyDAO

Added object of type NewUserCreatedObserver and passed the proxy down
object into its constructor so it could be added to the list of
observers.

Removed sysout message as it out put is being handled by newUserObserver
object as part of The Observer Pattern.